+++ app/src/main/java/com/vfgodoy/android_system_bar/view/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.vfgodoy.android_system_bar
+package com.vfgodoy.android_system_bar.view
 
 import android.os.Bundle
 import android.view.Menu
@@ -11,6 +11,7 @@ import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.appcompat.app.AppCompatActivity
+import com.vfgodoy.android_system_bar.R
 import com.vfgodoy.android_system_bar.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
@@ -36,7 +37,8 @@ class MainActivity : AppCompatActivity() {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top leve l destinations.
         appBarConfiguration = AppBarConfiguration(setOf(
-                R.id.nav_home, R.id.nav_products, R.id.nav_orders), drawerLayout)
+            R.id.nav_home, R.id.nav_products, R.id.nav_orders
+        ), drawerLayout)
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
     }
