+++ app/src/main/java/com/vfgodoy/android_system_bar/view/ProductsFragment.kt
@@ -1,4 +1,4 @@
-package com.vfgodoy.android_system_bar.ui.home
+package com.vfgodoy.android_system_bar.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -8,13 +8,13 @@ import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import com.vfgodoy.android_system_bar.R
-import com.vfgodoy.android_system_bar.databinding.FragmentHomeBinding
+import com.vfgodoy.android_system_bar.databinding.FragmentProductsBinding
+import com.vfgodoy.android_system_bar.viewmodel.ProductsViewModel
 
-class HomeFragment : Fragment() {
+class ProductsFragment : Fragment() {
 
-    private lateinit var homeViewModel: HomeViewModel
-    private var _binding: FragmentHomeBinding? = null
+    private lateinit var productsViewModel: ProductsViewModel
+    private var _binding: FragmentProductsBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -25,14 +25,14 @@ class HomeFragment : Fragment() {
             container: ViewGroup?,
             savedInstanceState: Bundle?
     ): View? {
-        homeViewModel =
-                ViewModelProvider(this).get(HomeViewModel::class.java)
+        productsViewModel =
+                ViewModelProvider(this).get(ProductsViewModel::class.java)
 
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentProductsBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        val textView: TextView = binding.textHome
-        homeViewModel.text.observe(viewLifecycleOwner, Observer {
+        val textView: TextView = binding.textProducts
+        productsViewModel.text.observe(viewLifecycleOwner, Observer {
             textView.text = it
         })
         return root
