+++ app/src/main/java/com/vfgodoy/android_system_bar/viewmodel/HomeViewModel.kt
@@ -1,4 +1,4 @@
-package com.vfgodoy.android_system_bar.ui.home
+package com.vfgodoy.android_system_bar.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
