+++ app/src/main/java/com/vfgodoy/android_system_bar/viewmodel/OrdersViewModel.kt
@@ -1,13 +1,13 @@
-package com.vfgodoy.android_system_bar.ui.products
+package com.vfgodoy.android_system_bar.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ProductsViewModel : ViewModel() {
+class OrdersViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is products Fragment"
+        value = "This is orders Fragment"
     }
     val text: LiveData<String> = _text
 }
