+++ app/src/main/res/layout/fragment_orders.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.orders.OrdersFragment">
+    tools:context=".view.OrdersFragment">
 
     <TextView
         android:id="@+id/text_orders"
