+++ app/src/main/res/layout/fragment_products.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.products.ProductsFragment">
+    tools:context=".view.ProductsFragment">
 
     <TextView
         android:id="@+id/text_products"
