+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.vfgodoy.android_system_bar.ui.home.HomeFragment"
+        android:name="com.vfgodoy.android_system_bar.view.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/nav_products"
-        android:name="com.vfgodoy.android_system_bar.ui.products.ProductsFragment"
+        android:name="com.vfgodoy.android_system_bar.view.ProductsFragment"
         android:label="@string/menu_products"
         tools:layout="@layout/fragment_products" />
 
     <fragment
         android:id="@+id/nav_orders"
-        android:name="com.vfgodoy.android_system_bar.ui.orders.OrdersFragment"
+        android:name="com.vfgodoy.android_system_bar.view.OrdersFragment"
         android:label="@string/menu_orders"
         tools:layout="@layout/fragment_orders" />
 </navigation>
